feat: testnet launch — fix 40+ TS errors, secure API routes, add templates, POUW infra
- Fix all TypeScript compile errors (target ES2022, type casts, missing fields)
- Secure build-provider and orchestration API routes with SECURITY_BEARER
- Add Cloudana deployment templates (WordPress, Node.js API, Python ML, etc.)
- Remove dead Akash template code
- Add POUW mining infrastructure (verifier, certificate store, rewards, chain recorder)
- Add provider hardware scan, faucet, auth middleware
- Add deployment infra (Dockerfiles, k8s manifests, subgraph)
- Add ProviderMinter and CLDFaucet contracts with tests
- Wire mining dashboard to live orchestrator API endpoints

+++ client/src/pages/workload-register.tsx
@@ -812,7 +812,7 @@ export default function WorkloadRegister() {
                     !fundWorkloadId ||
                     !fundAmount ||
                     isFundPending ||
-                    (cldAllowance != null && cldAllowance < parseEther(fundAmount || "0"))
+                    (cldAllowance != null && (cldAllowance as bigint) < parseEther(fundAmount || "0"))
                   }
                   onClick={() => {
                     fund(BigInt(fundWorkloadId), fundAmount || "0");
